Add tests for get_omim_number(hgnc_symbol)
